+++ docs/source/changelog.rst
@@ -9,6 +9,8 @@ Unreleased
 Version 0.133.1
 ===============
 
+* Fix cache of FCLCollisionManagerInstance for self-collision checker wasn't cleared for previously grabbed bodies even when they were removed from the env.
+
 * Fix initializing _listNonCollidingLinksWhenGrabbed based on incorrect grabbed bodies when grabbed bodies are shuffled between creation of Grabbed and Grabbed::ComputeListNonCollidingLinks.
 
 Version 0.133.0